services/relay/requirements.txt
# Relay-API Service Runtime Dependencies
# Phase 1: Foundation + Crypto

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# HTTP Client (inter-service communication)
httpx>=0.25.0

# End-to-end encryption (X25519 + AES-256-GCM via libsodium)
PyNaCl>=1.5.0

# Async support
anyio>=4.0.0

# Prometheus metrics
prometheus-client>=0.19.0

# ClamAV integration (optional)
pyclamd>=0.4.0

# Logging
python-json-logger>=2.0.0

# Configuration
python-dotenv>=1.0.0

# UUID v7 (time-sortable identifiers)
uuid6>=2.0.0

# Rate limiting (optional — graceful degradation when unavailable)
redis[hiredis]>=5.0.0

# PostgreSQL (batch result persistence)
asyncpg>=0.29.0
